refactor(agent-runtime): own prompt cache and registry facts
