+++ src/crates/core/src/agentic/agents/registry/custom.rs
@@ -9,7 +9,9 @@ use super::{AgentRegistry, CustomSubagentDetail};
 use crate::agentic::agents::definitions::custom::{CustomSubagent, CustomSubagentKind};
 use crate::agentic::agents::registry::types::subagent_key_for;
 use crate::agentic::agents::registry::visibility::SubagentVisibilityPolicy;
-use crate::agentic::agents::{Agent, AgentCategory, CustomSubagentConfig, SubAgentSource};
+use crate::agentic::agents::{
+    subagent_source_from_custom_kind, Agent, AgentCategory, CustomSubagentConfig, SubAgentSource,
+};
 use crate::agentic::tools::{get_all_registered_tool_names, get_readonly_registered_tool_names};
 use crate::service::config::global::GlobalConfigManager;
 use crate::service::config::mode_config_canonicalizer::persist_agent_profile_from_value;
@@ -37,7 +39,7 @@ impl AgentRegistry {
         let mut project_entries = HashMap::new();
         for mut sub in custom {
             let id = sub.id().to_string();
-            let source = SubAgentSource::from_custom_kind(sub.kind);
+            let source = subagent_source_from_custom_kind(sub.kind);
             // validate and correct tools and model
             Self::validate_custom_subagent(&mut sub, &valid_tools, &readonly_tools, &valid_models);
             // create CustomSubagentConfig cache configuration information
