+++ src/crates/core/src/agentic/agents/registry/query.rs
@@ -6,28 +6,20 @@ use super::support::{
 use super::AgentRegistry;
 use crate::agentic::agents::registry::types::{is_review_agent_entry, AgentEntry};
 use crate::agentic::agents::{
-    resolve_mode_config_profile_id, AgentCategory, AgentInfo, AgentToolPolicy, SubagentListScope,
-    SubagentQueryContext,
+    mode_presentation_rank, resolve_mode_config_profile_id, AgentCategory, AgentInfo,
+    AgentToolPolicy, SubagentListScope, SubagentQueryContext,
 };
 use crate::agentic::tools::get_all_registered_tool_names;
 use crate::service::config::mode_config_canonicalizer::resolve_effective_tools;
+use bitfun_agent_runtime::agents::subagent_source_presentation_rank;
 use std::collections::HashSet;
 use std::path::Path;
 
 impl AgentRegistry {
-    fn subagent_source_rank(source: Option<crate::agentic::agents::SubAgentSource>) -> u8 {
-        match source {
-            Some(crate::agentic::agents::SubAgentSource::Builtin) => 0,
-            Some(crate::agentic::agents::SubAgentSource::Project) => 1,
-            Some(crate::agentic::agents::SubAgentSource::User) => 2,
-            None => 3,
-        }
-    }
-
     fn sort_subagents_for_presentation(mut result: Vec<AgentInfo>) -> Vec<AgentInfo> {
         result.sort_by(|a, b| {
-            Self::subagent_source_rank(a.subagent_source)
-                .cmp(&Self::subagent_source_rank(b.subagent_source))
+            subagent_source_presentation_rank(a.subagent_source)
+                .cmp(&subagent_source_presentation_rank(b.subagent_source))
                 .then_with(|| a.id.to_lowercase().cmp(&b.id.to_lowercase()))
                 .then_with(|| a.id.cmp(&b.id))
                 .then_with(|| a.name.to_lowercase().cmp(&b.name.to_lowercase()))
@@ -113,21 +105,7 @@ impl AgentRegistry {
             .map(AgentInfo::from_agent_entry)
             .collect();
         drop(map);
-        result.sort_by(|a, b| {
-            let order = |id: &str| -> u8 {
-                match id {
-                    "agentic" => 0,
-                    "Cowork" => 1,
-                    "Plan" => 2,
-                    "debug" => 3,
-                    "Multitask" => 4,
-                    "DeepResearch" => 5,
-                    "Team" => 6,
-                    _ => 99,
-                }
-            };
-            order(&a.id).cmp(&order(&b.id))
-        });
+        result.sort_by(|a, b| mode_presentation_rank(&a.id).cmp(&mode_presentation_rank(&b.id)));
         result
     }
